requirements.txt
requests
python-dotenv
qdrant-client
FlagEmbedding
torch
